Add more metrics around indexing

+++ services/page-indexing-worker/src/main/kotlin/com/francisbailey/summitsearch/index/worker/task/PageIndexingTask.kt
@@ -12,6 +12,7 @@ import io.github.resilience4j.circuitbreaker.CircuitBreaker
 import io.github.resilience4j.ratelimiter.RateLimiterRegistry
 import io.micrometer.core.instrument.MeterRegistry
 import mu.KotlinLogging
+import java.lang.Exception
 import java.net.URL
 
 
@@ -45,7 +46,7 @@ class PageIndexingTask(
                 }
             }
         } else {
-            meterRegistry.counter("task.indexing.rate-limited").increment()
+            meterRegistry.counter("$TASK_METRIC.rate-limited").increment()
             log.warn { "Indexing rate exceeded for: $queueName. Backing off" }
         }
     }
@@ -55,7 +56,7 @@ class PageIndexingTask(
         log.info { "Found indexing task for: $queueName. Fetching Page: $pageUrl" }
 
         try {
-            val document = meterRegistry.timer("task.indexing.latency.page", "host", pageUrl.host).recordCallable {
+            val document = meterRegistry.timer("$TASK_METRIC.latency.page", "host", pageUrl.host).recordCallable {
                 pageCrawlerService.getHtmlDocument(pageUrl)
             }!!
 
@@ -63,7 +64,7 @@ class PageIndexingTask(
 
             dependencyCircuitBreaker.executeCallable {
                 linkDiscoveryService.submitDiscoveries(task, organicLinks)
-                meterRegistry.timer("task.indexing.indexservice.add.latency").recordCallable {
+                meterRegistry.timer("$TASK_METRIC.indexservice.add.latency").recordCallable {
                     indexService.indexPageContents(
                         SummitSearchIndexRequest(
                             source = pageUrl,
@@ -73,26 +74,41 @@ class PageIndexingTask(
                 }
 
                 log.info { "Successfully completed indexing task for: $queueName with $pageUrl" }
-                meterRegistry.counter("task.indexing.success", "host" , pageUrl.host).increment()
-                meterRegistry.counter("task.indexing.links.discovered").increment(organicLinks.size.toDouble())
+                meterRegistry.counter("$TASK_METRIC.success", "host" , pageUrl.host).increment()
+                meterRegistry.counter("$TASK_METRIC.links.discovered").increment(organicLinks.size.toDouble())
             }
         }
-        catch (e: RedirectedPageException) {
-            log.info { "Found page redirect. Submitting to discovery service if value is present" }
-            e.location?.run {
-                linkDiscoveryService.submitDiscoveries(task, listOf(this))
+        catch (e: Exception) {
+            meterRegistry.counter("$TASK_METRIC.exception.${e.cause}").increment()
+
+            when (e) {
+                is RedirectedPageException -> {
+                    log.info { "Found page redirect. Submitting to discovery service if value is present" }
+                    e.location?.run {
+                        meterRegistry.counter("$TASK_METRIC.redirects").increment()
+                        linkDiscoveryService.submitDiscoveries(task, listOf(this))
+                    }
+                }
+                is NonRetryablePageException -> {
+                    dependencyCircuitBreaker.executeCallable {
+                        log.error(e) { "Removing invalid content from index for: $pageUrl" }
+                        indexService.deletePageContents(SummitSearchDeleteIndexRequest(pageUrl))
+                    }
+                }
+                else -> {
+                    throw e
+                }
             }
         }
-        catch (e: NonRetryablePageException) {
-            log.error(e) { "Removing invalid content from index for: $pageUrl" }
-            dependencyCircuitBreaker.executeCallable {
-                indexService.deletePageContents(SummitSearchDeleteIndexRequest(pageUrl))
-            }
-        } finally {
+        finally {
             dependencyCircuitBreaker.executeCallable {
                 indexingTaskQueuePollingClient.deleteTask(task)
             }
         }
+    }
+
+    companion object {
+        const val TASK_METRIC = "task.indexing"
     }
 
 }
